feat: add root loader
